car_crosses_line now based on car's absolute pos
It works by simply comparing the car's worldPosition.x with TRACK_FINISH_LINE. This replaces the previous approach that was based on the calculated screen position. This is fraught with danger and is (probably) the cause for a bug where a car that didn't seem to finish "won" the race with a time of 0.000 seconds. This was on the phone, not the emulator.