Menu system working without memory leaks
(Apparently!) However, pressing BACK on the menu takes us to the game window, but does not exit the system. This is actually inconsistent between emulator and watch. On the emulator, the game quits after a few seconds, but this takes much longer (or maybe does not happen) on the watch.